refactor: separate editing of city and metadata
