+++ src/extensions/hooks/adopted-probe-city/package.json
@@ -37,6 +37,7 @@
 	"dependencies": {
 		"@directus/errors": "^0.2.0",
 		"@directus/extensions": "^0.1.0",
+		"@isaacs/ttlcache": "^1.4.1",
 		"any-ascii": "^0.3.2",
 		"axios": "^1.6.0"
 	}
